chore: repo cleanup — migrate Psirver threat model into README (#24)
* fix(psirver): bake the 1 GiB RLIMIT_FSIZE floor into Psirver's own default

clang++ on macOS writes large temporaries mid-compile, so Psirver's old
64 MiB RLIMIT_FSIZE default failed even a trivial C++ cell with "Filesize
limit exceeded". The dev launcher worked around this by exporting
PSIRVER_LIMIT_FSIZE_MB=1024, but that left the 64 MiB default lurking for
any other launch path (a hand-started Psirver per RUNNING.md, or a future
bundle). Move the safe value into Limits.cc so every launch path inherits
the headroom; the 10 s CPU and 15 s wall caps remain the real runaway
backstops, and operators running only trusted Python can still retighten
via the env var. Update dev.sh and RUNNING.md so the docs match.


* docs: pin a verified Gemma 4 E4B GGUF repo in getting-started

The Gemma download step used a <gemma-e4b-gguf-repo> placeholder, so a user
following getting-started hit unusable instructions on their first setup step.
Pin the llama.cpp team's instruction-tuned GGUF (ggml-org/gemma-4-E4B-it-GGUF,
Q4_K_M, ~5.3 GB), which matches the gemma-4-e4b alias config.py expects and
works with the documented `llama-server -hf` path.


* fix(psirver): accept ports above 32767 in select_port

select_port parsed the CLI port into a signed 16-bit temp (max 32767), so a
high port such as 50000 wrapped to a negative value and was rejected by the
"<= 0" guard — high ports were unusable. Parse into a full int, range-check
1..65535, then narrow to uint16_t, and return uint16_t to match the socket
API and the call site.


* docs: refresh status, roadmap, and known limitations for v1.0

Re-checked the README against what actually shipped through Phases 1-8:
Socratic mode, voice dictation, notebook add-cell/reload, and the timeline
replay scrubber are all in, so the stale "Not started / Stub / Pending" rows
and the "finish the Psirver job system" roadmap are replaced with the real
shipped set (functional core, execution hardening, restricted CORS, CI,
packaging) and a genuine future/stretch list.

Add a Known limitations section that names the edges v1.0 leaves for later
rather than hiding them: co-pilot mode resets to Direct on full reload, the
integrity marker is best-effort on a small model, the breadcrumb is a cosmetic
scratchpad.cpp label, the bundle ships the UI shell only, and the model-quality
numbers are still provisional. Fix the now-stale "/ai/transcribe is a stub"
notes in RUNNING.md and SMOKE_TEST.md (voice landed in #19), the bump the status
badge to v1.0, and correct the repo-layout's "stub/placeholder" annotations.


* chore(release): bump version to 1.0.0 and add CHANGELOG

Move every version marker from 0.1.0 to 1.0.0 — the FastAPI app, the backend
package, the desktop package, the Tauri bundle (and Cargo.lock to match) — so a
rebuilt bundle carries 1.0.0 (the .dmg name changes accordingly). Add a
CHANGELOG documenting the v1.0 feature set, hardening, tooling, the select_port
fix, and the known limitations.


* docs: add Apache-2.0 LICENSE and reference it from the README

Adds the verbatim Apache License 2.0 and replaces the README's "To be
determined" with a reference to it. Notes that services/psirver/ originated as
an Operating Systems course project and is included under the same license.


* docs: fold Psirver threat model into README, drop service-local README

Migrate the Phase 5 v1.0 security documentation out of
services/psirver/README.md into a new "Code execution & sandboxing"
section in the main README (after Architecture), preserving the threat
model and its scoping boundary, the PSIRVER_LIMIT_* configuration table,
the macOS RLIMIT_AS no-op caveat, and the v1.0 out-of-scope list. Add a
cross-link from SRS NFR-SEC-1 to the new section, then remove the now
fully-migrated services/psirver/README.md.


* chore: gitignore internal build-scaffolding docs

Guard against accidentally committing the external working docs
(PRODUCTION_PLAN.md, whetstone-issues.md, whetstone_prompts.md) that
were never tracked and should not ship in the portfolio repo.


---------

+++ services/psirver/src/utils.cc
@@ -25,9 +25,13 @@ void usage(const char* prog)
 // Parse command line arguments to find the port number. Library
 // functions used:
 // - std::stoi()
-int16_t select_port(int argc, char **argv)
+uint16_t select_port(int argc, char **argv)
 {
-  int16_t server_port_tmp = DEFAULT_PORT;
+  // Hold the parsed value in a full int: a port is 1-65535, which overflows a
+  // signed 16-bit temp (max 32767), so a high port like 50000 would wrap to a
+  // negative value and be rejected (or silently mis-bound). The range check
+  // below then validates it before the narrowing cast to uint16_t.
+  int server_port_tmp = DEFAULT_PORT;
   if (argc > 2) {		// Too many parameters
     usage(argv[0]);		// No return
   }
